+++ src/app/CameraController.h
@@ -10,7 +10,7 @@ public:
 	CameraController();
 	virtual ~CameraController();
 
-	void update(Camera & camera, GLfloat a_dt);
+	void update(Camera & camera, GLdouble a_dt);
 
 	void keyPress(Camera & a_camera, GLfloat a_dt) override;
 	void mouseMovement(Camera & a_camera) override;
